rename item-renderer.testdata
